+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+* The `--encoder="proto:<jq like expression>"` is deprecated, replaced with `--encoder="protojson:<jq like expression>"` (the accepted `jq` expression is very limited, see [README](./README.md) for details).
+
+* The `--encoder` flags is now fully optional and defaults to `parquet` (as it's the one best working with arbitrary protobuf).
+
+* Added Parquet output support through `--encoder=parquet`, see [README](./README.md) for full details of Parquet mode and how it works.
+
 ## v2.1.0
 
-* Fix last file erased when restarting from a cursor  
+* Fix last file erased when restarting from a cursor
 * Bumped substreams to v1.7.3
 * Enable gzip compression on substreams data on the wire